requirements.txt

python-telegram-bot==20.6
feedparser==6.0.10